requirements.txt
openai
langchain
langchain-openai
langchain-community
streamlit
langchain-experimental
supabase
python-dotenv